fix: 修复所有测试错误
- gateway/grpc.go: 修复 ConvertFromGatewayResponse nil pointer 解引用
- gateway/pool.go: 添加 activeCount 跟踪活跃连接数，修复 Get() 超限检查
- gateway/cache_test.go: 修复 int64 类型比较
- gateway/pool_test.go: 修复连接池测试逻辑，正确处理连接复用
- security/sensitive.go: 修复 MaskEmail 短 local 部分掩码
- security/security_test.go: 修复 SanitizeHTML 测试期望（/ 转义为 &#x2F;）
- metrics/metrics_test.go: 修复 int64 类型比较

+++ modules/gateway/cache_test.go
@@ -171,8 +171,8 @@ func TestLRUCache_Stats(t *testing.T) {
 	t.Run("initial stats", func(t *testing.T) {
 		stats := cache.Stats()
 		testhelpers.AssertEqual(t, stats.Entries, 0)
-		testhelpers.AssertEqual(t, stats.Hits, 0)
-		testhelpers.AssertEqual(t, stats.Misses, 0)
+		testhelpers.AssertEqual(t, stats.Hits, int64(0))
+		testhelpers.AssertEqual(t, stats.Misses, int64(0))
 		testhelpers.AssertEqual(t, stats.HitRate, 0.0)
 	})
 
@@ -190,8 +190,8 @@ func TestLRUCache_Stats(t *testing.T) {
 
 		stats := cache.Stats()
 		testhelpers.AssertEqual(t, stats.Entries, 2)
-		testhelpers.AssertEqual(t, stats.Hits, 2)
-		testhelpers.AssertEqual(t, stats.Misses, 2)
+		testhelpers.AssertEqual(t, stats.Hits, int64(2))
+		testhelpers.AssertEqual(t, stats.Misses, int64(2))
 		testhelpers.AssertEqual(t, stats.HitRate, 0.5)
 	})
 }
